requirements.txt
asyncio~=3.4.3
tornado==6.1
pyee~=9.0.4
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-cov==3.0.0
python-dotenv==0.20.0
pytest-mock==3.8.1
pyzmq==23.2.0
websockets==10.3
zmq==0.0.0